requirements.txt
python-chess==1.999
PyQt5==5.15.11
matplotlib==3.8.4
